build: update dependencies
Update workspace dependencies to latest versions:
- anyhow 1.0.99 -> 1.0.100
- clap 4.5.47 -> 4.5.53
- log 0.4.28 -> 0.4.29
- open 5.3.2 -> 5.3.3
- toml 0.9.5 -> 0.9.10
- rusqlite 0.37.0 -> 0.38.0
- jiff 0.2.15 -> 0.2.17

Fix time value cast to i64 in hotkey_manager.rs to match expected type.

+++ src/hotkey_manager.rs
@@ -115,7 +115,7 @@ impl HotKeyManager {
                                             SystemTime::now()
                                                 .duration_since(UNIX_EPOCH)
                                                 .unwrap() // should always success
-                                                .as_secs(),
+                                                .as_secs() as i64,
                                             path.to_string_lossy(),
                                         ),
                                     )
